output % kmers present / absent for read set on each node

+++ src/sglib/KmerCompressionIndex.hpp
@@ -24,7 +24,7 @@ public:
 
     void dump_histogram(std::string filename, uint16_t dataset=0);
 
-    double compute_compression_for_node(sgNodeID_t node, uint16_t max_graph_freq=10, uint16_t dataset=0);
+    double compute_compression_for_node(sgNodeID_t node, uint16_t max_graph_freq=10, int dataset=0);
     SequenceGraph & sg;
     std::vector<KmerCount> graph_kmers;
     std::vector<std::vector<uint16_t>> read_counts;
